@@ -1,6 +1,6 @@
 package edu.cs4730.flappyalien;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 /*
