@@ -1,7 +1,7 @@
 package edu.cs4730.textureviewdemo;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.Gravity;
 import android.widget.FrameLayout;
 
